{"schema":"decodex/commit/1","summary":"Consolidate tiny helper modules","authority":"manual"}
