--- apps/decodex/src/orchestrator/operator_http/assets/branding.rs
@@ -1,6 +1,0 @@
-pub(in crate::orchestrator::operator_http) const OPERATOR_DASHBOARD_ICON_PNG: &[u8] =
-	include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/orchestrator/assets/icon.png"));
-pub(in crate::orchestrator::operator_http) const OPERATOR_DASHBOARD_LOGO_ICO: &[u8] =
-	include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/orchestrator/assets/logo.ico"));
-pub(in crate::orchestrator::operator_http) const OPERATOR_DASHBOARD_LOGO_TOUCH_PNG: &[u8] =
-	include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/orchestrator/assets/logo-touch.png"));
