--- apps/decodex/src/orchestrator/status/autonomy/evidence/replay/completeness.rs
@@ -1,6 +1,0 @@
-pub(crate) fn operator_autonomy_evidence_completeness_rank(value: &str) -> u8 {
-	match value {
-		"complete" => 1,
-		_ => 0,
-	}
-}
