--- apps/decodex/src/recovery/git_worktree/command.rs
@@ -1,5 +1,0 @@
-use crate::prelude::Result;
-
-pub(in crate::recovery::git_worktree) fn trimmed_stdout(stdout: &[u8]) -> Result<String> {
-	Ok(String::from_utf8(stdout.to_vec())?.trim().to_owned())
-}
