+++ apps/decodex/src/recovery/stale_active_reentry.rs
@@ -2,12 +2,14 @@
 
 mod apply;
 mod control;
-mod evidence;
 mod local_cleanup;
 mod startable_restore;
 mod types;
 
 pub(super) use self::{
-	apply::apply_stale_active_release_reentries, evidence::evidence_contains,
-	types::StaleActiveReleaseReentryInput,
+	apply::apply_stale_active_release_reentries, types::StaleActiveReleaseReentryInput,
 };
+
+pub(in crate::recovery) fn evidence_contains(evidence: &[String], expected: &str) -> bool {
+	evidence.iter().any(|entry| entry == expected)
+}
