--- apps/decodex/src/recovery/stale_active_reentry/evidence.rs
@@ -1,3 +1,0 @@
-pub(in crate::recovery) fn evidence_contains(evidence: &[String], expected: &str) -> bool {
-	evidence.iter().any(|entry| entry == expected)
-}
