+++ apps/decodex/src/recovery/stale_active_reentry/local_cleanup.rs
@@ -1,7 +1,7 @@
 use crate::{
 	recovery::{
 		GHOST_LANE_TERMINAL_STATUS,
-		stale_active_reentry::{control, evidence, types::StaleActiveLocalCleanupReentryInput},
+		stale_active_reentry::{control, types::StaleActiveLocalCleanupReentryInput},
 	},
 	state::ProjectRunStatus,
 };
@@ -49,13 +49,13 @@ fn stale_active_local_cleanup_reentry_allowed(
 		&& control::reentry_control_channel_inactive_or_absent(
 			input.control_channel,
 			audit_evidence,
-		) && evidence::evidence_contains(
+		) && super::evidence_contains(
 		audit_evidence,
 		"only_stale_active_or_failed_control_evidence_present",
-	) && evidence::evidence_contains(audit_evidence, "review_lineage_missing")
-		&& evidence::evidence_contains(audit_evidence, "stale_active_release_audit_present")
-		&& evidence::evidence_contains(audit_evidence, "worktree_mapping_missing")
-		&& evidence::evidence_contains(audit_evidence, "worktree_missing")
+	) && super::evidence_contains(audit_evidence, "review_lineage_missing")
+		&& super::evidence_contains(audit_evidence, "stale_active_release_audit_present")
+		&& super::evidence_contains(audit_evidence, "worktree_mapping_missing")
+		&& super::evidence_contains(audit_evidence, "worktree_missing")
 }
 
 fn stale_active_local_cleanup_reentry_blocker(blocker: &str) -> bool {
