--- apps/radar/src/config.rs
@@ -1,3 +1,0 @@
-//! Checked-in Radar configuration locations.
-
-pub(crate) const DEFAULT_CONFIG_PATH: &str = "automations/radar/radar.toml";
